Split into client and scrobbler
